% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cityscapes-pix2pix.R
\name{cityscapes_pix2pix_dataset}
\alias{cityscapes_pix2pix_dataset}
\title{Cityscapes Pix2Pix dataset}
\usage{
cityscapes_pix2pix_dataset(
  root,
  split = "train",
  download = FALSE,
  ...,
  transform = NULL,
  target_transform = NULL
)
}
\arguments{
\item{root}{path to the data location}

\item{split}{train, test or valid}

\item{download}{wether to download or not}

\item{...}{Currently unused.}

\item{transform}{A function/transform that takes in an PIL image and returns
a transformed version. E.g, \code{\link[torchvision:transform_random_crop]{transform_random_crop()}}.}

\item{target_transform}{A function/transform that takes in the target and
transforms it.}
}
\description{
Downloads and prepares the cityscapes dataset that has been used in the
\href{https://arxiv.org/abs/1611.07004}{pix2pix paper}.
}
\details{
Find more information in the \href{https://phillipi.github.io/pix2pix/}{project website}
}
