% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess-the-correlation.R
\name{guess_the_correlation_dataset}
\alias{guess_the_correlation_dataset}
\title{Guess The Correlation dataset}
\usage{
guess_the_correlation_dataset(
  root,
  token = NULL,
  split = "train",
  transform = NULL,
  indexes = NULL,
  download = FALSE
)
}
\arguments{
\item{root}{path to the data location}

\item{token}{a path to the json file obtained in Kaggle. See \href{https://pins.rstudio.com/articles/boards-kaggle.html}{here}
for additional info.}

\item{split}{string. 'train' or 'submition'}

\item{transform}{function that receives a torch tensor and return another torch tensor, transformed.}

\item{indexes}{set of integers for subsampling (e.g. 1:140000)}

\item{download}{wether to download or not}
}
\description{
Prepares the Guess The Correlation dataset available in Kaggle \href{https://www.kaggle.com/c/guess-the-correlation}{here}
}
\details{
We use pins for downloading and managing authetication.
If you want to download the dataset you need to register the Kaggle board as
described in \href{https://pins.rstudio.com/articles/boards-kaggle.html}{this link}.
or pass the \code{token} argument.
}
