
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spam-dataloader.R
\name{spam_dataloader}
\alias{spam_dataloader}
\title{Spam Data Loader}
\usage{
spam_dataloader(
  url = "https://hastie.su.domains/ElemStatLearn/datasets/spam.data",
  batch_size = 32,
  shuffle = TRUE,
  download = FALSE
)
}
\arguments{
\item{url}{A character string representing the URL of the dataset.}

\item{batch_size}{Number of samples per batch. Defaults to 32.}

\item{shuffle}{Logical; whether to shuffle the data. Defaults to TRUE.}

\item{download}{Logical; whether to download the dataset. Defaults to FALSE.}
}
\value{
A torch dataloader object for the spam dataset.
}
\description{
A dataloader for the spam dataset commonly used in machine learning.
}
\examples{
\dontrun{
# Simple usage:
loader <- spam_dataloader(download = TRUE)
batch <- dataloader_make_iter(loader) \%>\% dataloader_next()
dim(batch$x)
length(batch$y)
}
}
