% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spam-dataloader.R
\name{spam_dataset}
\alias{spam_dataset}
\title{Spam Dataset Loader}
\usage{
spam_dataset(
  url = "https://hastie.su.domains/ElemStatLearn/datasets/spam.data",
  download = FALSE,
  transform = NULL,
  target_transform = NULL
)
}
\arguments{
\item{url}{A character string representing the URL of the dataset.}

\item{download}{Logical; whether to download the dataset. Defaults to FALSE.}

\item{transform}{Function to apply transformations to the features. Defaults to NULL.}

\item{target_transform}{Function to apply transformations to the labels. Defaults to NULL.}
}
\value{
A \code{torch::dataset} object for the spam dataset.
}
\description{
Defines the spam dataset commonly used in machine learning.
}
\examples{
\dontrun{
# Simple usage:
ds <- spam_dataset(download = TRUE)
loader <- dataloader(ds, batch_size = 32, shuffle = TRUE)
batch <- dataloader_make_iter(loader) \%>\% dataloader_next()
dim(batch$x)
length(batch$y)
}
}
